Stable version with all features

+++ server/src/columns/columns.service.ts
@@ -1,11 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { BaseColumnsDto } from './dtos/base.dto';
 import { ColumnsGateway } from './columns.gateway';
 import { MoveColumnDto } from './dtos/moveColumn.dto';
 import { TasksService } from 'src/tasks/tasks.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateColumnDto } from './dtos/createColumn.dto';
 import { RenameColumnDto } from './dtos/renameColumn.dto';
+import { DeleteColumnDto } from './dtos/deleteColumn.dto';
+
+const defaultColumnNames = ['to do', 'doing', 'done'];
 
 @Injectable()
 export class ColumnsService {
@@ -20,7 +22,10 @@ export class ColumnsService {
         const isColumnNameTaken = !!(await this.prismaService.column.findFirst({
             where: {
                 AND: {
-                    name: body.name,
+                    name: {
+                        equals: body.name.trim(),
+                        mode: 'insensitive',
+                    },
                     boardId: body.boardData.id,
                 },
             },
@@ -36,16 +41,12 @@ export class ColumnsService {
             },
         });
 
-        const newColumnPosition =
-            allColumns.sort((a, b) => b.position - a.position)[0].position + 1;
-
         //create the column
-        //set column position == columns + 1
-        await this.prismaService.column.create({
+        const { id } = await this.prismaService.column.create({
             data: {
                 name: body.name,
                 boardId: body.boardData.id,
-                position: newColumnPosition,
+                position: allColumns.length,
             },
         });
 
@@ -55,11 +56,26 @@ export class ColumnsService {
             message: 'New column added.',
             affectedBoardId: body.boardData.id,
         });
+
+        return id;
     }
 
-    async delete(body: BaseColumnsDto) {
-        //the deletion includes the task steps
+    async delete(body: DeleteColumnDto) {
+        if (
+            defaultColumnNames.includes(
+                body.columnData.name.trim().toLowerCase(),
+            )
+        ) {
+            throw new Error('You cannot delete the default table columns!');
+        }
+
+        //the task deletion is cascading
         await this.tasksService.deleteMany(body.columnData.id);
+        await this.prismaService.column.delete({
+            where: {
+                id: body.columnData.id,
+            },
+        });
     }
 
     async deleteMany(boardId: number) {
@@ -71,7 +87,7 @@ export class ColumnsService {
 
         //deletes all tasks and steps cascadingly
         await Promise.all(
-            columns.map((column) => async () => {
+            columns.map(async (column) => {
                 await this.tasksService.deleteMany(column.id);
             }),
         );
@@ -83,73 +99,16 @@ export class ColumnsService {
         });
     }
 
-    async changePosition(body: MoveColumnDto) {
-        //check if the new position is valid
-        const allBoardColumns = await this.prismaService.column.findMany({
-            where: {
-                AND: [{ boardId: body.boardData.id }],
-            },
-        });
-
-        // its is not possible to have empty array as at least 3 columns will be created by default and will not be erasable
-        const maxPosition = allBoardColumns
-            .map((column) => column.position)
-            .sort((a, b) => b - a)[0];
-
-        if (body.newPosition < 0 || body.newPosition > maxPosition) {
-            throw new Error('Invalid new position!');
-        }
-
-        // Update the column's position
-        await this.prismaService.column.update({
-            where: { id: body.columnId },
-            data: {
-                position: body.newPosition,
-            },
-        });
-
-        // Update the positions of other columns if needed
-        if (body.columnData.position < body.newPosition) {
-            // Shift columns between the old and new positions up by 1
-            await this.prismaService.column.updateMany({
-                where: {
-                    AND: [
-                        { boardId: body.columnData.boardId },
-                        {
-                            position: {
-                                gte: body.columnData.position,
-                                lte: body.newPosition - 1,
-                            },
-                        },
-                    ],
-                },
-                data: {
-                    position: { increment: 1 },
-                },
-            });
-        } else if (body.columnData.position > body.newPosition) {
-            // Shift columns between the new and old positions down by 1
-            await this.prismaService.column.updateMany({
-                where: {
-                    AND: [
-                        { boardId: body.columnData.boardId },
-                        {
-                            position: {
-                                gte: body.newPosition,
-                                lte: body.columnData.position - 1,
-                            },
-                        },
-                    ],
-                },
-                data: {
-                    position: { increment: -1 },
-                },
-            });
-        }
-    }
-
     async rename(body: RenameColumnDto) {
-        const isColumnNameTaken = await this.prismaService.column.findFirst({
+        if (
+            defaultColumnNames.includes(
+                body.columnData.name.trim().toLowerCase(),
+            )
+        ) {
+            throw new Error('You cannot rename the default table columns!');
+        }
+
+        const columnsWithSameName = await this.prismaService.column.findMany({
             where: {
                 AND: {
                     name: body.newName,
@@ -157,6 +116,13 @@ export class ColumnsService {
                 },
             },
         });
+
+        const isColumnNameTaken =
+            columnsWithSameName.filter(
+                (column) =>
+                    column.name === body.newName &&
+                    column.id !== body.columnData.id,
+            ).length > 0;
 
         if (isColumnNameTaken) {
             throw new Error('Column name is taken!');
@@ -168,6 +134,107 @@ export class ColumnsService {
             },
             data: {
                 name: body.newName,
+            },
+        });
+    }
+
+    async move(body: MoveColumnDto) {
+        const columnsInsideBoard = await this.prismaService.column.findMany({
+            where: {
+                boardId: body.boardData.id,
+            },
+        });
+
+        if (body.destinationPosition == body.columnData.position) {
+            return;
+        }
+
+        // Ensure the destination position is within the valid range
+        if (body.destinationPosition >= columnsInsideBoard.length) {
+            body.destinationPosition = columnsInsideBoard.length - 1;
+        }
+
+        if (body.destinationPosition > body.columnData.position) {
+            const matches = await this.prismaService.column.findMany({
+                where: {
+                    AND: [
+                        {
+                            position: {
+                                gt: body.columnData.position,
+                            },
+                        },
+                        {
+                            position: {
+                                lte: body.destinationPosition,
+                            },
+                        },
+                        { boardId: body.boardData.id },
+                    ],
+                },
+            });
+
+            await Promise.all(
+                matches.map(async (column) => {
+                    await this.prismaService.column.update({
+                        where: {
+                            id: column.id,
+                        },
+                        data: {
+                            position: column.position - 1,
+                        },
+                    });
+                }),
+            );
+
+            await this.prismaService.column.update({
+                where: {
+                    id: body.columnData.id,
+                },
+                data: {
+                    position: body.destinationPosition,
+                },
+            });
+
+            return;
+        }
+
+        const matches = await this.prismaService.column.findMany({
+            where: {
+                AND: [
+                    {
+                        position: {
+                            gte: body.destinationPosition,
+                        },
+                    },
+                    {
+                        position: {
+                            lt: body.columnData.position,
+                        },
+                    },
+                    { boardId: body.boardData.id },
+                ],
+            },
+        });
+
+        await Promise.all(
+            matches.map(async (column) => {
+                await this.prismaService.column.update({
+                    where: {
+                        id: column.id,
+                    },
+                    data: {
+                        position: column.position + 1,
+                    },
+                });
+            }),
+        );
+
+        await this.prismaService.column.update({
+            where: {
+                id: body.columnData.id,
+            },
+            data: {
+                position: body.destinationPosition,
             },
         });
     }
